Create How To Install R and rstudio on Ubuntu 16.04
R is a popular open source programming language that specializes in statistical computing and graphics. It is widely used for developing statistical software and performing data analysis.
RStudio is a set of integrated tools designed to help you be more productive with R. It includes a console, syntax-highlighting editor that supports direct code execution, and a variety of robust tools for plotting, viewing history, debugging and managing your workspace. Learn More about RStudio features.